TM_NsのPML粗方完成
HにdNSをつける方を採用（PML有・無とも）
若干の誤差が残るのは問題点
MurはSt用を周期境界に代用することで実装可能
こちらも誤差あり

+++ Field.cpp
@@ -1,6 +1,6 @@
 #include"Field.h"
 
-Field::Field(int width, int height, int h_u, int pml)
+Field::Field(int width, int height, double h_u, int pml)
 :WIDTH(width),
 HEIGHT(height),
 H_U(h_u),
